pyproject.toml
[project]
name = "coscientist"
version = "0.1.0"
description = "Atomic skills for a personal academic research agent in Claude Code"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Coscientist" }]
dependencies = [
    # Core
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "python-slugify>=8.0",
    # Arxiv fast-path
    "arxiv2markdown>=0.1",
    # PDF extraction
    "docling>=2.0",
    "pymupdf>=1.24",
    # Browser automation for institutional-access
    "playwright>=1.48",
    # Misc
    "rich>=13.7",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6",
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]
ignore = ["E501"]
